rename: move-menu to legacy-menu
